Move MethodBody into own schema
