+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/Enhancements.scala
@@ -13,16 +13,22 @@ object Enhancements extends SchemaBase {
       | Note: these should *NOT* be written by the language frontend.
       |""".stripMargin
 
-  def apply(builder: SchemaBuilder, base: Base.Schema, methodSchema: Method.Schema, typeDeclSchema: TypeDecl.Schema) =
-    new Schema(builder, base, methodSchema, typeDeclSchema)
+  def apply(builder: SchemaBuilder,
+            base: Base.Schema,
+            methodSchema: Method.Schema,
+            methodBody: MethodBody.Schema,
+            typeDeclSchema: TypeDecl.Schema) =
+    new Schema(builder, base, methodSchema, methodBody, typeDeclSchema)
 
   class Schema(builder: SchemaBuilder,
                base: Base.Schema,
                methodSchema: Method.Schema,
+               methodBody: MethodBody.Schema,
                typeDeclSchema: TypeDecl.Schema) {
 
     import base._
     import methodSchema._
+    import methodBody._
     import typeDeclSchema._
     implicit private val schemaInfo = SchemaInfo.forClass(getClass)
