+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/MetaData.scala
@@ -1,11 +1,12 @@
 package io.shiftleft.codepropertygraph.schema
 
-import overflowdb.schema.{Cardinality, NodeType, SchemaBuilder, SchemaInfo}
+import overflowdb.schema.{Cardinality, Constant, NodeType, SchemaBuilder, SchemaInfo}
 import overflowdb.storage.ValueTypes
 
 object MetaData extends SchemaBase {
 
   def index: Int = 1
+  override def providedByFrontend: Boolean = true
 
   override def description: String =
     """
@@ -46,6 +47,20 @@ object MetaData extends SchemaBase {
       )
       .protoId(39)
       .addProperties(language, version, overlays, hash)
+
+    val languages = builder.addConstants(
+      category = "Languages",
+      Constant(name = "JAVA", value = "JAVA", valueType = ValueTypes.STRING, comment = "").protoId(1),
+      Constant(name = "JAVASCRIPT", value = "JAVASCRIPT", valueType = ValueTypes.STRING, comment = "").protoId(2),
+      Constant(name = "GOLANG", value = "GOLANG", valueType = ValueTypes.STRING, comment = "").protoId(3),
+      Constant(name = "CSHARP", value = "CSHARP", valueType = ValueTypes.STRING, comment = "").protoId(4),
+      Constant(name = "C", value = "C", valueType = ValueTypes.STRING, comment = "").protoId(5),
+      Constant(name = "PYTHON", value = "PYTHON", valueType = ValueTypes.STRING, comment = "").protoId(6),
+      Constant(name = "LLVM", value = "LLVM", valueType = ValueTypes.STRING, comment = "").protoId(7),
+      Constant(name = "PHP", value = "PHP", valueType = ValueTypes.STRING, comment = "").protoId(8),
+      Constant(name = "FUZZY_TEST_LANG", value = "FUZZY_TEST_LANG", valueType = ValueTypes.STRING, comment = "")
+        .protoId(9),
+    )
 
   }
