+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/SchemaBase.scala
@@ -3,4 +3,5 @@ package io.shiftleft.codepropertygraph.schema
 trait SchemaBase {
   def index: Int
   def description: String
+  def providedByFrontend: Boolean = false
 }
